テキスト追加

+++ Parupino/Source/CharaData/Command.cpp
@@ -32,7 +32,7 @@ namespace GAME
 		UINT frame = ( m_limitTime < len ) ? m_limitTime : len;
 		
 		//完成入力数（カウントダウン）(判定は-1まで用いる)
-		int comp = m_vecGameKey.size() - 1;
+		int comp = (int)m_vecGameKey.size() - 1;
 
 		//フレームを遡って検索
 		for ( UINT i = 0; i < frame; ++i )
